requirements.txt
torch>=2.2
torchvision>=0.17
facenet-pytorch>=2.5.3
numpy>=1.26
pillow>=10.0
scikit-learn>=1.4
pandas>=2.2
matplotlib>=3.8
flask>=3.0
python-dotenv>=1.0
joblib>=1.3
